Added character and enemy
Utilised character and enemy classes, displaying it on the LCD and removed the object that was mimicking a character from the Map.cpp.  Now the enemy spawns at a random position on the path and the character starts from the beginning of the path. Also increased the map size and path width a bit.