Started working on the combat
Created a combat system class, to implement features for the fighting phase. For now, there is a combat area indicator that appears when the character is within a specific range of the enemy, to indicate the area where you can fight. Also added health bars for the character and enemy.