Implemented winning and losing
Now if the timer runs out before you find the enemies and kill them you will get a screen showing defeat but if you kill the enemies before the timer runs out you win. Additionally, added a power up button which restarts the timer, however, it can only be used once in the whole game, so you have to choose the correct time to activate it.